chore: reduce exported-symbol footprint
